footballtalks/requirements.txt
streamlit
requests
pandas
numpy
scikit-learn
matplotlib
plotly
python-dotenv
